refactor(ui): modular components, pages shell, agents subdirs
- Organize feature UI under components/{chat,channels,agents,tasks}
- Nest agents/profile and agents/activity under agents
- Move Sidebar, MainPanel, TabBar to pages/
- Remove unused prototype and demo components
- Normalize channels directory to lowercase for cross-platform paths
- Update AGENTS.md project layout

Made-with: Cursor
