+++ ui/src/App.tsx
@@ -1,8 +1,8 @@
 import { QueryClientProvider } from '@tanstack/react-query'
 import { queryClient } from './queryClient'
 import { AppProvider } from './store'
-import { Sidebar } from './components/Sidebar'
-import { MainPanel } from './components/MainPanel'
+import { Sidebar } from './pages/Sidebar'
+import { MainPanel } from './pages/MainPanel'
 
 export default function App() {
   return (
