+++ ui/src/components/agents/WorkspacePanel.css
