+++ ui/src/components/agents/activity/ActivityPanel.css
