+++ ui/src/components/chat/ChatPanel.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useRef } from 'react'
 import { Search, Settings2, Users } from 'lucide-react'
-import { useApp } from '../store'
+import { useApp } from '../../store'
 import { MessageItem } from './MessageItem'
-import type { HistoryMessage } from '../types'
+import type { HistoryMessage } from '../../types'
 import { useVisibilityTracking } from '@/hooks/useVisibilityTracking'
 import './ChatPanel.css'
