+++ ui/src/components/chat/MessageInput.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState, useRef } from 'react'
 import { Paperclip, Plus } from 'lucide-react'
-import { useApp } from '../store'
-import { useHistory } from '../hooks/useHistory'
-import { sendMessage, createTasks, uploadFile } from '../api'
+import { useApp } from '../../store'
+import { useHistory } from '../../hooks/useHistory'
+import { sendMessage, createTasks, uploadFile } from '../../api'
 import { MentionTextarea } from './MentionTextarea'
 import type { MentionMember } from './MentionTextarea'
 import { ToastRegion } from './ToastRegion'
