+++ ui/src/pages/MainPanel.tsx
@@ -3,17 +3,17 @@ import { getChannelMembers, getTeam, sendMessage } from '../api'
 import { useApp, useTarget } from '../store'
 import { useHistory } from '../hooks/useHistory'
 import { TabBar } from './TabBar'
-import { ChatHeader, ChatPanel } from './ChatPanel'
-import { TasksPanel } from './TasksPanel'
-import { ProfilePanel } from './ProfilePanel'
-import { ActivityPanel } from './ActivityPanel'
-import { WorkspacePanel } from './WorkspacePanel'
-import { MessageInput } from './MessageInput'
-import { ThreadPanel } from './ThreadPanel'
-import { ThreadsTab } from './ThreadsTab'
-import { ChannelMembersPanel } from './ChannelMembersPanel'
+import { ChatHeader, ChatPanel } from '../components/chat/ChatPanel'
+import { TasksPanel } from '../components/tasks/TasksPanel'
+import { ProfilePanel } from '../components/agents/profile/ProfilePanel'
+import { ActivityPanel } from '../components/agents/activity/ActivityPanel'
+import { WorkspacePanel } from '../components/agents/WorkspacePanel'
+import { MessageInput } from '../components/chat/MessageInput'
+import { ThreadPanel } from '../components/chat/ThreadPanel'
+import { ThreadsTab } from '../components/chat/ThreadsTab'
+import { ChannelMembersPanel } from '../components/channels/ChannelMembersPanel'
 import type { ChannelMemberInfo, TeamResponse } from '../types'
-import { TeamSettings } from './TeamSettings'
+import { TeamSettings } from '../components/channels/TeamSettings'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 
 export function MainPanel() {
